Created LargeTitle component and loginStyles
LargeTitle component is to be used for headers in each login screen.

loginStyles are styles that would apply to each screen during the login process. An example is done on the LandingPage.js for how the views and styles are to be applied. The main idea is to align top & center, flex whatever space is leftover in the middle and align center the bottom buttons after that.

Also did slight modifications to LargeButton to accommodate the new loginStyles