Added LineDivider Component
LineDivider Component added for easy use and standard styling.

Changed other components to have a prop called "topSpacing" default space above elements is 15, but this prop can change this easily so we have the option to modify this or add views (with specified heights) to vertically space elements.

Also added height prop in LargeTitle component