Added UserPrompt Component
UserPrompt component is just a centered view with text to greet and guide the user upon each screen.

Other minor changes: I added default values to all my props in components. For instance, topSpacing default is 15. If no value or a value of 0 is passed, 15 will be used. To pass a value of 0, pass -1 instead. Any other value will override it properly.