fix: PhoneNumber error
We found that we previously stored the phoneNumber from auth to the database. This is different from the exact text string we receive on input from the login button press.

To fix this we stuck with the phone number formatting from auth (same characters, just no spaces). When querying the database, we now successfully pull users with the corresponding phone number.

Co-Authored-By: Evan C. Navarro <EvanCNavarro@gmail.com>