Hiding Avatar for multiple messages
When multiple messages are sent by the same user, the avatar is no longer shown. The message is displayed right under the previous message. Messages from new users has their avatar.

This is done by creating a map from the current message to the previous one. That map is used to check if the phone number of the previous message is the same or different and renders based on that.