Attempted GroupMembers
Tried to get the Users with the sam ids as the group, but was only partially successful

I got the array of user ids from the group just fine, but couldn't figure out how to pull the users with the same ids afterwards.

The UI elements are setup so that once that happens, it'll display a list of users once the code is uncommented.